Revert incompatible scaling changes
This reverts commits 5eee6d925b91b447b415e58d50b138a8a098daec
and 7a83376c3a58a157f0230e7b82384bbb5134a5b6 from
https://github.com/pinterf/xy-VSFilter/pull/24 and its follow-up fix.

The dependence on the videos storage resolution described in
https://github.com/pinterf/xy-VSFilter/issues/17 is actually
part of the ASS format and required to be kept for compatibility with
existing files.
Keeping compatible with this beaviour is also why the xy-SubFilter
interface in include/SubRenderIntf.h takes an "originalVideoSize"
parameter; as unlike xy-VSFilter, xy-SubFilter can render at native
display resolution. (While for xy-VSFilter, rendering always happens at
the videos original storage resolution, so the rendering dimension
already carries the informaton and a separate parameter is not needed).